Move more Typescript code to 'common'
