+++ html-report/src/components/common/TestExecution.ts
@@ -1,4 +1,4 @@
-import Selection from "./Selection";
+import Selection from "./Selection.ts";
 
 export default class TestExecution {
