+++ html-report/src/components/common/TestResultStatus.ts
