+++ html-report/src/main.ts
@@ -3,7 +3,7 @@ import {createI18n} from 'vue-i18n'
 
 import './style.css'
 import App from './App.vue'
-import TestExecution from './TestExecution'
+import TestExecution from './components/common/TestExecution.ts'
 import VueEasyLightbox from "vue-easy-lightbox";
 
 const app = createApp(App, {
